Add additional info to keys
Started saving keys as tuples so it's easier to link the retrieved key to the other data structures referencing it in the program (particularly the blackboard artifacts).

Reworked the processing to accommodate this.  Switched from looping based on retrieved values to looping based on provided tuples and just handling the exception if it pops up.  Slightly more verbose but significantly less complicated.