Rework reg key classification
Get the registry key type from the rejistry library, which is nicer and easier to maintain than keeping the type in the key tuple.

Also attempt to decode everything as a date, then give up easily if it doesn't work.  Not great, but the best workaround possible since for reasons known only to our lord and savior Jesus Christ some timestamp registry keys are stored as strings and others are stored as numbers.